backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
google-genai>=1.68.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
rembg[cpu]>=2.0.0
pillow>=10.0.0
